вивільнила головний index.js
