+++ src/js/refs.js
@@ -1,9 +1,5 @@
 export default {
-
-
   modalCloseBtn: document.querySelector('button[data-modal-close]'),
-
-
     cardContainer: document.querySelector('.main-section-event-list'),
     chooseCountryInput: document.querySelector('.choose-country'),
     searchingInput: document.querySelector('.form-field'),
